requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Speech-to-Text (OpenAI Whisper)
# Whisper (requires FFmpeg installed on the system):
openai-whisper
ffmpeg-python

# Text-to-Speech (Amazon Polly)
boto3==1.34.0
botocore==1.34.0

# NLP (Hugging Face)
transformers==4.35.0
torch>=2.1.0
sentencepiece==0.1.99

# Google Gemini API (for Response Generation)
google-generativeai>=0.3.0

# Audio processing
pyaudio==0.2.13
wave==0.0.2

# Backend/Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0

# Web Framework for Dashboard
flask==3.0.0
flask-cors==4.0.0

# Utilities
requests==2.31.0
numpy>=1.25.2,<2.0.0

